requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
openai==1.44.0
networkx==3.3
GitPython==3.1.43
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
redis==5.0.8
pymysql==1.1.1
cryptography==43.0.1
SQLAlchemy==2.0.34
slowapi==0.1.9
python-json-logger==2.0.7
httpx==0.27.2
pytest==8.3.3
